move tests closer to the source, set up e2e tests
