+++ src/lib/__tests__/errors.test.ts
