+++ src/lib/__tests__/hash.test.ts
