+++ src/services/__tests__/registry.test.ts
