+++ tsconfig.build.json
@@ -9,7 +9,9 @@
     "jest.config.ts",
     "dist",
     "node_modules",
-    "tests"
+    "**/__tests__/**/*",
+    "tests",
+    "e2e"
   ],
   "tsc-alias": {
     "verbose": true,
