+++ tsconfig.json
@@ -50,14 +50,16 @@
       "@lib/*": [
         "src/lib/*"
       ],
-      "@mocks/*": [
-        "tests/mocks/*"
+      "@tests/*": [
+        "tests/*"
       ]
     }
   },
   "include": [
     "src/**/*",
     "tests/**/*",
+    "e2e/**/*",
+    "**/__tests__/**/*",
     "tsconfig.json",
     "jest.setup.ts",
     "jest.config.ts",
